rename views to ui
